LPS-194405 Move the configuration code to layout-locked-layouts-web
