+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/init.jsp
@@ -23,7 +23,6 @@ taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 <%@ page import="com.liferay.layout.content.page.editor.web.internal.constants.ContentPageEditorWebKeys" %><%@
 page import="com.liferay.layout.content.page.editor.web.internal.display.context.ContentPageEditorDisplayContext" %><%@
 page import="com.liferay.layout.content.page.editor.web.internal.display.context.EditCollectionConfigurationDisplayContext" %><%@
-page import="com.liferay.layout.content.page.editor.web.internal.display.context.LockedLayoutsConfigurationDisplayContext" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
