+++ modules/apps/layout/layout-locked-layouts-web/src/main/java/com/liferay/layout/locked/layouts/web/internal/display/context/LockedLayoutsConfigurationDisplayContext.java
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.content.page.editor.web.internal.display.context;
+package com.liferay.layout.locked.layouts.web.internal.display.context;
 
-import com.liferay.layout.content.page.editor.web.internal.configuration.LockedLayoutsConfiguration;
+import com.liferay.layout.locked.layouts.web.internal.configuration.LockedLayoutsConfiguration;
 
 /**
  * @author Lourdes Fernández Besada
