+++ modules/apps/layout/layout-locked-layouts-web/src/main/resources/META-INF/resources/configuration/locked_layouts.jsp
